feat: move types to components
