+++ src/handlers/index.ts
@@ -1,2 +1,2 @@
 export * from "./factory";
-export type * from "./types";
+export type * from "../components/types";
